+++ NEWS.md
@@ -1,3 +1,5 @@
+# rlang (development version)
+
 # rlang 1.1.4
 
 * Added missing C level `r_dyn_raw_push_back()` and `r_dyn_chr_push_back()`